Abstract gcs publisher interface to wipe those if-else off
