+++ src/ray/gcs/gcs_server/gcs_job_manager.h
@@ -18,7 +18,7 @@
 #include "ray/gcs/gcs_server/gcs_function_manager.h"
 #include "ray/gcs/gcs_server/gcs_init_data.h"
 #include "ray/gcs/gcs_server/gcs_table_storage.h"
-#include "ray/gcs/pubsub/gcs_pub_sub.h"
+#include "ray/gcs/pubsub/gcs_publisher.h"
 #include "ray/rpc/gcs_server/gcs_rpc_server.h"
 
 namespace ray {
