+++ src/ray/gcs/gcs_server/gcs_worker_manager.h
@@ -15,7 +15,7 @@
 #pragma once
 
 #include "ray/gcs/gcs_server/gcs_table_storage.h"
-#include "ray/gcs/pubsub/gcs_pub_sub.h"
+#include "ray/gcs/pubsub/gcs_publisher.h"
 #include "ray/rpc/gcs_server/gcs_rpc_server.h"
 
 namespace ray {
